refactor [#490] 유저 온보딩 상태 조회 API URI 변경
* refactor [#490] 유저 온보딩 상태 조회 API URI 변경

* refactor [#484] 온보딩 과정에서 중복된 아티스트가 응답되지 않도록 개선

* refactor [#484] 상위 아티스트 목록 조회 API에서 조회된 상위 아티스트 목록을 유저에 캐싱

* refactor [#484] 특정 아티스트와 관련된 아티스트 목록 조회 API에서 중복된 아티스트를 응답하지 않도록 수정

* refactor [#484] 아티스트 연관 검색어 목록 조회 API에서 이미 추가된 아티스트 연관 검색어 제외

* refactor [#484] 연관 검색 결과의 특정 아티스트 추가 (추적용) API 구현

* refactor [#484] 온보딩 API이 정상적으로 수행될 경우 캐시된 아티스트 데이터 삭제

* refactor [#484] Redis TTL 1 day로 설정

* feat [#479] 온보딩 여부 조회 API 구현

* feat [#479] 온보딩 여부 조회 API 구현

* docs [#479] openapi3.yaml 파일이 포함되지 않도록 gitignore 수정

* refactor [#482] RestAssured 테스트 개선

* refactor [#482] 태그 추가를 위한 문서 스니펫 추가

* refactor [#482] 테스트에 사용된 인증 헤더 값 숨김 처리

* refactor [#482] 필드 값에 개행을 추가할 수 있도록 수정

* refactor [#482] APIBaseTest 변경사항 반영

* refactor [#482] APIBaseTest 변경사항 반영

* feat [#479] 온보딩 여부 조회 API 구현

* feat [#479] 온보딩 여부 조회 API 구현

* docs [#479] openapi3.yaml 파일이 포함되지 않도록 gitignore 수정

* refactor [#482] modifyUri로 수정한 값이 openapi3.yaml에 반영되지 않아 제거

* etc [#482] 이거 왜 자꾸 생기지

* refactor [#482] 태그 추가를 위한 문서 스니펫 추가

* refactor [#482] 테스트에 사용된 인증 헤더 값 숨김 처리

* refactor [#482] 필드 값에 개행을 추가할 수 있도록 수정

* refactor [#482] APIBaseTest 변경사항 반영

* refactor [#482] APIBaseTest 변경사항 반영

* refactor [#482] modifyUri로 수정한 값이 openapi3.yaml에 반영되지 않아 제거

* feat [#487] 홈 통합 검색 API - 검색어 기준에 검색어 저장 추가

* feat [#487] 홈 통합 검색 API - 검색어 기준에 검색어 저장 추가

* refactor [#482] RestAssured 테스트 개선

* refactor [#482] 태그 추가를 위한 문서 스니펫 추가

* refactor [#482] 테스트에 사용된 인증 헤더 값 숨김 처리

* refactor [#482] 필드 값에 개행을 추가할 수 있도록 수정

* refactor [#482] APIBaseTest 변경사항 반영

* refactor [#482] APIBaseTest 변경사항 반영

* feat [#479] 온보딩 여부 조회 API 구현

* feat [#479] 온보딩 여부 조회 API 구현

* docs [#479] openapi3.yaml 파일이 포함되지 않도록 gitignore 수정

* refactor [#482] modifyUri로 수정한 값이 openapi3.yaml에 반영되지 않아 제거

* etc [#482] 이거 왜 자꾸 생기지

* refactor [#482] 태그 추가를 위한 문서 스니펫 추가

* refactor [#482] 테스트에 사용된 인증 헤더 값 숨김 처리

* refactor [#482] 필드 값에 개행을 추가할 수 있도록 수정

* refactor [#482] APIBaseTest 변경사항 반영

* refactor [#482] APIBaseTest 변경사항 반영

* refactor [#482] modifyUri로 수정한 값이 openapi3.yaml에 반영되지 않아 제거

* feat [#487] 홈 통합 검색 API - 검색어 기준에 검색어 저장 추가

* refactor [#484] 불필요한 트랜잭션 제거

* refactor [#484] 상위 아티스트 목록 조회 API에서 조회된 상위 아티스트 목록을 유저에 캐싱

* refactor [#484] 특정 아티스트와 관련된 아티스트 목록 조회 API에서 중복된 아티스트를 응답하지 않도록 수정

* refactor [#484] 아티스트 연관 검색어 목록 조회 API에서 이미 추가된 아티스트 연관 검색어 제외

* refactor [#484] 연관 검색 결과의 특정 아티스트 추가 (추적용) API 구현

* refactor [#484] 온보딩 API이 정상적으로 수행될 경우 캐시된 아티스트 데이터 삭제

* refactor [#484] Redis TTL 1 day로 설정

* refactor [#484] 불필요한 트랜잭션 제거

* refactor [#490] 유저 온보딩 상태 조회 API URI 변경
