+++ src/main/java/org/sopt/confeti/api/user/controller/UserOnboardController.java
@@ -4,9 +4,11 @@ import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
 import lombok.RequiredArgsConstructor;
 import org.sopt.confeti.api.user.dto.response.UserOnboardTopArtistsResponse;
+import org.sopt.confeti.api.user.dto.response.onboard.GetIsOnboardingResponse;
 import org.sopt.confeti.api.user.dto.response.onboard.UserOnboardRelatedArtistsResponse;
 import org.sopt.confeti.api.user.facade.UserOnboardFacade;
 import org.sopt.confeti.api.user.facade.dto.response.UserOnboardTopArtistsDTO;
+import org.sopt.confeti.api.user.facade.dto.response.onboard.GetIsOnboardingDTO;
 import org.sopt.confeti.api.user.facade.dto.response.onboard.UserOnboardRelatedArtistsDTO;
 import org.sopt.confeti.domain.user.constant.Role;
 import org.sopt.confeti.global.annotation.Permission;
@@ -73,5 +75,12 @@ public class UserOnboardController {
     ) {
         userOnboardFacade.cacheTopArtist(userId, artistId);
         return ApiResponseUtil.success(SuccessMessage.SUCCESS);
+    }
+
+    @Permission(role = {Role.ONBOARDING, Role.GENERAL, Role.ADMIN})
+    @GetMapping("/state")
+    public ResponseEntity<BaseResponse<?>> getIsOnboarding(@UserId Long userId) {
+        GetIsOnboardingDTO isOnboardingDTO = userOnboardFacade.getIsOnboarding(userId);
+        return ApiResponseUtil.success(SuccessMessage.SUCCESS, GetIsOnboardingResponse.from(isOnboardingDTO));
     }
 }
