+++ src/main/java/org/sopt/confeti/api/user/dto/response/onboard/GetIsOnboardingResponse.java
@@ -1,6 +1,6 @@
-package org.sopt.confeti.api.auth.dto.response;
+package org.sopt.confeti.api.user.dto.response.onboard;
 
-import org.sopt.confeti.api.auth.facade.dto.response.GetIsOnboardingDTO;
+import org.sopt.confeti.api.user.facade.dto.response.onboard.GetIsOnboardingDTO;
 
 public record GetIsOnboardingResponse(
         boolean isOnboarding
