+++ src/main/java/org/sopt/confeti/api/user/facade/UserOnboardFacade.java
@@ -5,9 +5,11 @@ import java.util.Set;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.sopt.confeti.api.user.facade.dto.response.UserOnboardTopArtistsDTO;
+import org.sopt.confeti.api.user.facade.dto.response.onboard.GetIsOnboardingDTO;
 import org.sopt.confeti.api.user.facade.dto.response.onboard.UserOnboardRelatedArtistsDTO;
-import org.sopt.confeti.domain.user.application.UserOnboardService;
 import org.sopt.confeti.domain.user.application.UserService;
+import org.sopt.confeti.domain.user.constant.Role;
+import org.sopt.confeti.domain.user.application.UserOnboardService;
 import org.sopt.confeti.domain.user.application.dto.request.UserOnboardCacheTopArtistsDTO;
 import org.sopt.confeti.global.annotation.Facade;
 import org.sopt.confeti.global.exception.NotFoundException;
@@ -16,6 +18,7 @@ import org.sopt.confeti.global.resolver.music_api.artist.vo.ConfetiArtist;
 import org.sopt.confeti.global.resolver.music_api.music.vo.ConfetiMusic;
 import org.sopt.confeti.global.resolver.music_api.music.vo.ConfetiMusicArtist;
 import org.sopt.confeti.global.util.music.MusicAPIHandler;
+import org.springframework.transaction.annotation.Transactional;
 
 @Facade
 @RequiredArgsConstructor
@@ -91,5 +94,11 @@ public class UserOnboardFacade {
         topArtistIds.addAll(artistIds);
 
         userOnboardService.cacheTopArtists(userId, UserOnboardCacheTopArtistsDTO.from(topArtistIds));
+    }
+
+    @Transactional(readOnly = true)
+    public GetIsOnboardingDTO getIsOnboarding(long userId) {
+        Role userRole = userService.getRole(userId);
+        return GetIsOnboardingDTO.from(userRole);
     }
 }
